Add Open Graph meta tags and og-image.png
Added Open Graph meta tags to the index.html file to improve SEO and social media sharing. These include title, description, and image properties for better representation when the site link is shared. An og-image.png file was added to the public directory to be used as a thumbnail image for shared links.