Remove deprecated server:start script
Removed the "server:start" script from the package.json file as it is no longer needed. This script was used for starting a JSON server, which has since been deprecated in favor of a new backend service. The change helps in maintaining a clean and up-to-date list of scripts in the project, reducing potential confusion for developers.